try to replace uint32 to int64

+++ be/src/vec/common/format_ip.h
@@ -117,10 +117,10 @@ inline void formatIPv4(const unsigned char* src, char*& dst, uint8_t mask_tail_o
  */
 template <typename T, typename EOFfunction>
     requires(std::is_same<typename std::remove_cv<T>::type, char>::value)
-inline bool parseIPv4(T*& src, EOFfunction eof, unsigned char* dst, int32_t first_octet = -1) {
+inline bool parseIPv4(T*& src, EOFfunction eof, unsigned char* dst, int64_t first_octet = -1) {
     if (src == nullptr || first_octet > 255) return false;
 
-    uint32_t result = 0;
+    int64_t result = 0;
     int offset = 24;
     if (first_octet >= 0) {
         result |= first_octet << offset;
@@ -130,7 +130,7 @@ inline bool parseIPv4(T*& src, EOFfunction eof, unsigned char* dst, int32_t firs
     for (; true; offset -= 8, ++src) {
         if (eof()) return false;
 
-        uint32_t value = 0;
+        int64_t value = 0;
         size_t len = 0;
         while (is_numeric_ascii(*src) && len <= 3) {
             value = value * 10 + (*src - '0');
